move MaterialChipsInput to org.sufficientlysecure namespace
