--- extern/MaterialChipsInput/src/main/java/com/pchmn/materialchips/RecyclerItemClickListener.java
@@ -1,51 +1,0 @@
-package com.pchmn.materialchips;
-
-import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.view.GestureDetector;
-import android.view.MotionEvent;
-import android.view.View;
-
-public class RecyclerItemClickListener implements RecyclerView.OnItemTouchListener {
-    private OnItemClickListener mListener;
-    private boolean mIgnoreTouch = false;
-
-    public interface OnItemClickListener {
-        void onItemClick(View view, int position);
-    }
-
-    private GestureDetector mGestureDetector;
-
-    public RecyclerItemClickListener(Context context, OnItemClickListener listener) {
-        mListener = listener;
-        mGestureDetector = new GestureDetector(context, new GestureDetector.SimpleOnGestureListener() {
-            @Override
-            public boolean onSingleTapUp(MotionEvent e) {
-                return true;
-            }
-        });
-    }
-
-    @Override
-    public boolean onInterceptTouchEvent(RecyclerView view, MotionEvent e) {
-        if (mIgnoreTouch) {
-            return false;
-        }
-        View childView = view.findChildViewUnder(e.getX(), e.getY());
-        if (childView != null && mListener != null && mGestureDetector.onTouchEvent(e)) {
-            mListener.onItemClick(childView, view.getChildAdapterPosition(childView));
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public void onTouchEvent(RecyclerView view, MotionEvent motionEvent) {
-        // TODO: should we move mListener.onItemClick here
-    }
-
-    @Override
-    public void onRequestDisallowInterceptTouchEvent(boolean disallowIntercept) {
-        mIgnoreTouch = disallowIntercept;
-    }
-}
