--- extern/MaterialChipsInput/src/main/java/com/pchmn/materialchips/model/ChipInterface.java
@@ -1,10 +1,0 @@
-package com.pchmn.materialchips.model;
-
-
-import com.pchmn.materialchips.adapter.FilterableAdapter.FilterableItem;
-
-
-public interface ChipInterface extends FilterableItem {
-    String getLabel();
-    String getInfo();
-}
