+++ extern/MaterialChipsInput/src/main/java/org/sufficientlysecure/materialchips/model/ChipInterface.java
@@ -1,0 +1,10 @@
+package org.sufficientlysecure.materialchips.model;
+
+
+import org.sufficientlysecure.materialchips.adapter.FilterableAdapter.FilterableItem;
+
+
+public interface ChipInterface extends FilterableItem {
+    String getLabel();
+    String getInfo();
+}
