+++ extern/MaterialChipsInput/src/main/res/layout/chips_input.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.pchmn.materialchips.views.ScrollViewMaxHeight xmlns:android="http://schemas.android.com/apk/res/android"
+<org.sufficientlysecure.materialchips.views.ScrollViewMaxHeight xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/container"
     android:layout_width="match_parent"
@@ -14,4 +14,4 @@
         android:paddingTop="8dp"
         android:paddingBottom="8dp" />
 
-</com.pchmn.materialchips.views.ScrollViewMaxHeight>
+</org.sufficientlysecure.materialchips.views.ScrollViewMaxHeight>
