+++ extern/MaterialChipsInput/src/main/res/values/colors.xml
@@ -5,6 +5,5 @@
     <color name="chips_opened_bg">#009688</color>
     <color name="colorCloseChip">#ababab</color>
     <color name="white_transparency">#b9ffffff</color>
-    <color name="colorAccent">?attr/colorAccent</color>
 
 </resources>
